// src/app/models/noticia.model.ts
// Generated by https://quicktype.io

export interface Noticia {
  idNoticia:            number;
  titulo:               string;
  contenido:            string;
  linkNoticia:          string;
  latitud:              number;
  longitud:             number;
  urlImagen:            string;
  fileBytes:            null;
  idProveedor:          number;
  enviado:              boolean;
  usuarioCreacion:      number;
  fechaCreacion:        string;
  fechaCreacionUtc:     string;
  usuarioModificacion:  number;
  fechaModificacion:    string;
  fechaModificacionUtc: string;
  proveedor:            null;
  idProveedorWeb:       number;
  idUsuarioWeb:         number;
  proveedorWeb:         null;
}